requirements.txt
Django
django-jet
djangorestframework
dj-database-url
gunicorn
psycopg2-binary
unidecode
whitenoise
django-cors-headers
google-auth
requests
beautifulsoup4